Move spgrep.pointgroup
