+++ src/spgrep/irreps.py
@@ -9,7 +9,6 @@ from warnings import warn
 import numpy as np
 
 from spgrep._constants import ATOL, MAX_NUM_RANDOM_GENERATIONS, RTOL
-from spgrep.pointgroup import get_pointgroup_chain_generators
 from spgrep.rep.group import get_identity_index, get_inverse_index, get_order
 from spgrep.rep.irreps import frobenius_schur_indicator
 from spgrep.rep.representation import get_character, get_intertwiner
@@ -17,6 +16,7 @@ from spgrep.symmetry.group import (
     get_cayley_table,
     get_factor_system_from_little_group,
 )
+from spgrep.symmetry.pointgroup import get_pointgroup_chain_generators
 from spgrep.symmetry.representation import get_projective_regular_representation
 from spgrep.utils import NDArrayComplex, NDArrayFloat, NDArrayInt, nroot
