+++ src/spgrep/symmetry/pointgroup.py
@@ -1,0 +1,718 @@
+"""Decomposition of crystallographic point groups as solvable groups."""
+
+from __future__ import annotations
+
+import numpy as np
+from spglib import get_pointgroup
+
+from spgrep.utils import NDArrayInt, ndarray2d_to_integer_tuple
+
+# List of point groups after applying transformation matrices given by `spglib.get_pointgroup`
+# See https://github.com/spglib/spglib/issues/164
+# Operations are ordered as same as Table 3.2.3.2 of ITA (2016).
+# fmt: off
+pg_dataset = {
+    "1": [[((1, 0, 0), (0, 1, 0), (0, 0, 1))]],  # 0: 1
+    "-1": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 1: -1
+        ]
+    ],
+    "2": [
+        # unique axis-a
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),
+        ],
+        # unique axis-b
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 1: 2
+        ],
+        # unique axis-c
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),
+        ],
+    ],
+    "m": [
+        # unique axis-a
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),
+        ],
+        # unique axis-b
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: m
+        ],
+        # unique axis-c
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),
+        ],
+    ],
+    "2/m": [
+        # unique axis-b
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 1: 2
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 2: -1
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 3: m
+        ]
+    ],
+    "222": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 3: 2_100
+        ]
+    ],
+    "mm2": [
+        # unique axis-a
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),
+        ],
+        # unique axis-b
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),
+        ],
+        # unique axis-c
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),   # 2: m_010
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),   # 3: m_100
+        ],
+    ],
+    "mmm": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   # 1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),   # 2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),   # 3: 2_100
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 4: -1
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    # 5: m_001
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),    # 6: m_010
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 7: m_100
+        ]
+    ],
+    "4": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),   # 2: 4^+
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),   # 3: 4^-
+        ]
+    ],
+    "-4": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),  # 2: -4^+
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),  # 3: -4^-
+        ]
+    ],
+    "4/m": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   # 1: 2
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),    # 2: 4^+
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),    # 3: 4^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 4: -1
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    # 5: m
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),   # 6: -4^+
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),   # 7: -4^-
+        ]
+    ],
+    "422": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   # 1: 2_001
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),    # 2: 4^+
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),    # 3: 4^-
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),   # 4: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),   # 5: 2_100
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    # 6: 2_110
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  # 7: 2_1-10
+        ]
+    ],
+    "4mm": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),   # 2: 4^+
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),   # 3: 4^-
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),   # 4: m_010
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),   # 5: m_100
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),  # 6: m_110
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),    # 7: m_1-10
+        ]
+    ],
+    "-42m": [
+        # -42m
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2_001
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),  # 2: -4^+
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),  # 3: -4^-
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 4: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 5: 2_100
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),  # 6: m_110
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),    # 7: m_1-10
+        ],
+        # -4m2
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),
+        ],
+    ],
+    "4/mmm": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  1: 2_001
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),    #  2: 4^+
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),    #  3: 4^-
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),   #  4: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),   #  5: 2_100
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    #  6: 2_110
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  #  7: 2_1-10
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  #  8: -1
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    #  9: m_001
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),   # 10: -4^+
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),   # 11: -4^-
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),    # 12: m_010
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 13: m_100
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),   # 14: m_110
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),     # 15: m_1-10
+        ]
+    ],
+    "3": [
+        # Hexagonal axes
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),  # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),  # 2: 3^-
+        ]
+    ],
+    "-3": [
+        # Hexagonal axes
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   # 2: 3^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 3: -1
+            ((0, 1, 0), (-1, 1, 0), (0, 0, -1)),   # 4: -3^+
+            ((1, -1, 0), (1, 0, 0), (0, 0, -1)),   # 5: -3^-
+        ]
+    ],
+    "32": [
+        # 312
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   # 2: 3^-
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  # 3: 2_1-10
+            ((-1, 1, 0), (0, 1, 0), (0, 0, -1)),   # 4: 2_120
+            ((1, 0, 0), (1, -1, 0), (0, 0, -1)),   # 5: 2_210
+        ],
+        # 321
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   # 2: 3^-
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    # 3: 2_110
+            ((1, -1, 0), (0, -1, 0), (0, 0, -1)),  # 4: 2_100
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, -1)),  # 5: 2_010
+        ],
+    ],
+    "3m": [
+        # 3m1
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),  # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),  # 2: 3^-
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),  # 3: m_110
+            ((-1, 1, 0), (0, 1, 0), (0, 0, 1)),   # 4: m_100
+            ((1, 0, 0), (1, -1, 0), (0, 0, 1)),   # 5: m_010
+        ],
+        # 31m
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),
+        ],
+    ],
+    "-3m": [
+        # -31m
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   #  2: 3^-
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  #  3: 2_1-10
+            ((-1, 1, 0), (0, 1, 0), (0, 0, -1)),   #  4: 2_120
+            ((1, 0, 0), (1, -1, 0), (0, 0, -1)),   #  5: 2_210
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  #  6: -1
+            ((0, 1, 0), (-1, 1, 0), (0, 0, -1)),   #  7: -3^+
+            ((1, -1, 0), (1, 0, 0), (0, 0, -1)),   #  8: -3^-
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),     #  9: m_1-10
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),   # 10: m_120
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),   # 11: m_210
+        ],
+        # -3m1
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   #  2: 3^-
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    #  3: 2_110
+            ((1, -1, 0), (0, -1, 0), (0, 0, -1)),  #  4: 2_100
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, -1)),  #  5: 2_010
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  #  6: -1
+            ((0, 1, 0), (-1, 1, 0), (0, 0, -1)),   #  7: -3^+
+            ((1, -1, 0), (1, 0, 0), (0, 0, -1)),   #  8: -3^-
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),   #  9: m_110
+            ((-1, 1, 0), (0, 1, 0), (0, 0, 1)),    # 10: m_100
+            ((1, 0, 0), (1, -1, 0), (0, 0, 1)),    # 11: m_010
+        ],
+    ],
+    "6": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),  # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),  # 2: 3^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 3: 2
+            ((0, 1, 0), (-1, 1, 0), (0, 0, 1)),   # 4: 6^-
+            ((1, -1, 0), (1, 0, 0), (0, 0, 1)),   # 5: 6^+
+        ]
+    ],
+    "-6": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     # 0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   # 1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   # 2: 3^-
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    # 3: m
+            ((0, -1, 0), (1, -1, 0), (0, 0, -1)),  # 4: -6^-
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, -1)),  # 5: -6^+
+        ]
+    ],
+    "6/m": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   #  2: 3^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  3: 2
+            ((0, 1, 0), (-1, 1, 0), (0, 0, 1)),    #  4: 6^-
+            ((1, -1, 0), (1, 0, 0), (0, 0, 1)),    #  5: 6^+
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  #  6: -1
+            ((0, 1, 0), (-1, 1, 0), (0, 0, -1)),   #  7: -3^+
+            ((1, -1, 0), (1, 0, 0), (0, 0, -1)),   #  8: -3^-
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    #  9: m
+            ((0, -1, 0), (1, -1, 0), (0, 0, -1)),  # 10: -6^-
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, -1)),  # 11: -6^+
+        ]
+    ],
+    "622": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   #  2: 3^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  3: 2_001
+            ((0, 1, 0), (-1, 1, 0), (0, 0, 1)),    #  4: 6^-
+            ((1, -1, 0), (1, 0, 0), (0, 0, 1)),    #  5: 6^+
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    #  6: 2_110
+            ((1, -1, 0), (0, -1, 0), (0, 0, -1)),  #  7: 2_100
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, -1)),  #  8: 2_010
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  #  9: 2_1-10
+            ((-1, 1, 0), (0, 1, 0), (0, 0, -1)),   # 10: 2_120
+            ((1, 0, 0), (1, -1, 0), (0, 0, -1)),   # 11: 2_210
+        ]
+    ],
+    "6mm": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),  #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),  #  2: 3^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  #  3: 2_001
+            ((0, 1, 0), (-1, 1, 0), (0, 0, 1)),   #  4: 6^-
+            ((1, -1, 0), (1, 0, 0), (0, 0, 1)),   #  5: 6^+
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),  #  6: m_110
+            ((-1, 1, 0), (0, 1, 0), (0, 0, 1)),   #  7: m_100
+            ((1, 0, 0), (1, -1, 0), (0, 0, 1)),   #  8: m_010
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),    #  9: m_1-10
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),  # 10: m_120
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),  # 11: m_210
+        ]
+    ],
+    "-6m2": [
+        # -6m2
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   #  2: 3^-
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    #  3: m_001
+            ((0, -1, 0), (1, -1, 0), (0, 0, -1)),  #  4: -6^-
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, -1)),  #  5: -6^+
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),   #  6: m_110
+            ((-1, 1, 0), (0, 1, 0), (0, 0, 1)),    #  7: m_100
+            ((1, 0, 0), (1, -1, 0), (0, 0, 1)),    #  8: m_010
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  #  9: 2_1-10
+            ((-1, 1, 0), (0, 1, 0), (0, 0, -1)),   # 10: 2_120
+            ((1, 0, 0), (1, -1, 0), (0, 0, -1)),   # 11: 2_210
+        ],
+        # -62m
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),
+            ((0, -1, 0), (1, -1, 0), (0, 0, -1)),
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, -1)),
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),
+            ((1, -1, 0), (0, -1, 0), (0, 0, -1)),
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, -1)),
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),
+        ],
+    ],
+    "6/mmm": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((0, -1, 0), (1, -1, 0), (0, 0, 1)),   #  1: 3^+
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, 1)),   #  2: 3^-
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  3: 2_001
+            ((0, 1, 0), (-1, 1, 0), (0, 0, 1)),    #  4: 6^-
+            ((1, -1, 0), (1, 0, 0), (0, 0, 1)),    #  5: 6^+
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    #  6: 2_110
+            ((1, -1, 0), (0, -1, 0), (0, 0, -1)),  #  7: 2_100
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, -1)),  #  8: 2_010
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  #  9: 2_1-10
+            ((-1, 1, 0), (0, 1, 0), (0, 0, -1)),   # 10: 2_120
+            ((1, 0, 0), (1, -1, 0), (0, 0, -1)),   # 11: 2_210
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 12: -1
+            ((0, 1, 0), (-1, 1, 0), (0, 0, -1)),   # 13: -3^+
+            ((1, -1, 0), (1, 0, 0), (0, 0, -1)),   # 14: -3^-
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    # 15: m_001
+            ((0, -1, 0), (1, -1, 0), (0, 0, -1)),  # 16: -6^-
+            ((-1, 1, 0), (-1, 0, 0), (0, 0, -1)),  # 17: -6^+
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),   # 18: m_110
+            ((-1, 1, 0), (0, 1, 0), (0, 0, 1)),    # 19: m_100
+            ((1, 0, 0), (1, -1, 0), (0, 0, 1)),    # 20: m_010
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),     # 21: m_1-10
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),   # 22: m_120
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),   # 23: m_210
+        ]
+    ],
+    "23": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 3: 2_100
+            ((0, 0, 1), (1, 0, 0), (0, 1, 0)),    # 4: 3^+_111
+            ((0, 0, 1), (-1, 0, 0), (0, -1, 0)),  # 5: 3^+_-11-1
+            ((0, 0, -1), (-1, 0, 0), (0, 1, 0)),  # 6: 3^+_1-1-1
+            ((0, 0, -1), (1, 0, 0), (0, -1, 0)),  # 7: 3^+_-1-11
+            ((0, 1, 0), (0, 0, 1), (1, 0, 0)),    # 8: 3^-_111
+            ((0, -1, 0), (0, 0, 1), (-1, 0, 0)),  # 9: 3^-_1-1-1
+            ((0, 1, 0), (0, 0, -1), (-1, 0, 0)),  # 10: 3^-_-1-11
+            ((0, -1, 0), (0, 0, -1), (1, 0, 0)),  # 11: 3^-_-11-1
+        ]
+    ],
+    "m-3": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),   #  2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),   #  3: 2_100
+            ((0, 0, 1), (1, 0, 0), (0, 1, 0)),     #  4: 3^+_111
+            ((0, 0, 1), (-1, 0, 0), (0, -1, 0)),   #  5: 3^+_-11-1
+            ((0, 0, -1), (-1, 0, 0), (0, 1, 0)),   #  6: 3^+_1-1-1
+            ((0, 0, -1), (1, 0, 0), (0, -1, 0)),   #  7: 3^+_-1-11
+            ((0, 1, 0), (0, 0, 1), (1, 0, 0)),     #  8: 3^-_111
+            ((0, -1, 0), (0, 0, 1), (-1, 0, 0)),   #  9: 3^-_1-1-1
+            ((0, 1, 0), (0, 0, -1), (-1, 0, 0)),   # 10: 3^-_-1-11
+            ((0, -1, 0), (0, 0, -1), (1, 0, 0)),   # 11: 3^-_-11-1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 12: -1
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    # 13: m_001
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),    # 14: m_010
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 15: m_100
+            ((0, 0, -1), (-1, 0, 0), (0, -1, 0)),  # 16: -3^+_111
+            ((0, 0, -1), (1, 0, 0), (0, 1, 0)),    # 17: -3^+_-11-1
+            ((0, 0, 1), (1, 0, 0), (0, -1, 0)),    # 18: -3^+_1-1-1
+            ((0, 0, 1), (-1, 0, 0), (0, 1, 0)),    # 19: -3^+_-1-11
+            ((0, -1, 0), (0, 0, -1), (-1, 0, 0)),  # 20: -3^-_111
+            ((0, 1, 0), (0, 0, -1), (1, 0, 0)),    # 21: -3^-_1-1-1
+            ((0, -1, 0), (0, 0, 1), (1, 0, 0)),    # 22: -3^-_-1-11
+            ((0, 1, 0), (0, 0, 1), (-1, 0, 0)),    # 23: -3^-11-1
+        ]
+    ],
+    "432": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),   #  2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),   #  3: 2_100
+            ((0, 0, 1), (1, 0, 0), (0, 1, 0)),     #  4: 3^+_111
+            ((0, 0, 1), (-1, 0, 0), (0, -1, 0)),   #  5: 3^+_-11-1
+            ((0, 0, -1), (-1, 0, 0), (0, 1, 0)),   #  6: 3^+_1-1-1
+            ((0, 0, -1), (1, 0, 0), (0, -1, 0)),   #  7: 3^+_-1-11
+            ((0, 1, 0), (0, 0, 1), (1, 0, 0)),     #  8: 3^-_111
+            ((0, -1, 0), (0, 0, 1), (-1, 0, 0)),   #  9: 3^-_1-1-1
+            ((0, 1, 0), (0, 0, -1), (-1, 0, 0)),   # 10: 3^-_-1-11
+            ((0, -1, 0), (0, 0, -1), (1, 0, 0)),   # 11: 3^-_-11-1
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    # 12: 2_110
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  # 13: 2_1-10
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),    # 14: 4^-_001
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),    # 15: 4^+_001
+            ((1, 0, 0), (0, 0, 1), (0, -1, 0)),    # 16: 4^-_100
+            ((-1, 0, 0), (0, 0, 1), (0, 1, 0)),    # 17: 2_011
+            ((-1, 0, 0), (0, 0, -1), (0, -1, 0)),  # 18: 2_01-1
+            ((1, 0, 0), (0, 0, -1), (0, 1, 0)),    # 19: 4^+_100
+            ((0, 0, 1), (0, 1, 0), (-1, 0, 0)),    # 20: 4^+_010
+            ((0, 0, 1), (0, -1, 0), (1, 0, 0)),    # 21: 2_101
+            ((0, 0, -1), (0, 1, 0), (1, 0, 0)),    # 22: 4^-_010
+            ((0, 0, -1), (0, -1, 0), (-1, 0, 0)),  # 23: 2_-101
+        ]
+    ],
+    "-43m": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),    #  0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  #  1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  #  2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),  #  3: 2_100
+            ((0, 0, 1), (1, 0, 0), (0, 1, 0)),    #  4: 3^+_111
+            ((0, 0, 1), (-1, 0, 0), (0, -1, 0)),  #  5: 3^+_-11-1
+            ((0, 0, -1), (-1, 0, 0), (0, 1, 0)),  #  6: 3^+_1-1-1
+            ((0, 0, -1), (1, 0, 0), (0, -1, 0)),  #  7: 3^+_-1-11
+            ((0, 1, 0), (0, 0, 1), (1, 0, 0)),    #  8: 3^-_111
+            ((0, -1, 0), (0, 0, 1), (-1, 0, 0)),  #  9: 3^-_1-1-1
+            ((0, 1, 0), (0, 0, -1), (-1, 0, 0)),  # 10: 3^-_-1-11
+            ((0, -1, 0), (0, 0, -1), (1, 0, 0)),  # 11: 3^-_-11-1
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),    # 12: m_1-10
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),  # 13: m_110
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),  # 14: -4^+_001
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),  # 15: -4^-_001
+            ((1, 0, 0), (0, 0, 1), (0, 1, 0)),    # 16: m_01-1
+            ((-1, 0, 0), (0, 0, 1), (0, -1, 0)),  # 17: -4^+_100
+            ((-1, 0, 0), (0, 0, -1), (0, 1, 0)),  # 18: -4^-_100
+            ((1, 0, 0), (0, 0, -1), (0, -1, 0)),  # 19: m_011
+            ((0, 0, 1), (0, 1, 0), (1, 0, 0)),    # 20: m_-101
+            ((0, 0, 1), (0, -1, 0), (-1, 0, 0)),  # 21: -4^-_010
+            ((0, 0, -1), (0, 1, 0), (-1, 0, 0)),  # 22: m_101
+            ((0, 0, -1), (0, -1, 0), (1, 0, 0)),  # 23: -4^+_010
+        ]
+    ],
+    "m-3m": [
+        [
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),     #  0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),   #  1: 2_001
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),   #  2: 2_010
+            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),   #  3: 2_100
+            ((0, 0, 1), (1, 0, 0), (0, 1, 0)),     #  4: 3^+_111
+            ((0, 0, 1), (-1, 0, 0), (0, -1, 0)),   #  5: 3^+_-11-1
+            ((0, 0, -1), (-1, 0, 0), (0, 1, 0)),   #  6: 3^+_1-1-1
+            ((0, 0, -1), (1, 0, 0), (0, -1, 0)),   #  7: 3^+_-1-11
+            ((0, 1, 0), (0, 0, 1), (1, 0, 0)),     #  8: 3^-_111
+            ((0, -1, 0), (0, 0, 1), (-1, 0, 0)),   #  9: 3^-_1-1-1
+            ((0, 1, 0), (0, 0, -1), (-1, 0, 0)),   # 10: 3^-_-1-11
+            ((0, -1, 0), (0, 0, -1), (1, 0, 0)),   # 11: 3^-_-11-1
+            ((0, 1, 0), (1, 0, 0), (0, 0, -1)),    # 12: 2_110
+            ((0, -1, 0), (-1, 0, 0), (0, 0, -1)),  # 13: 2_1-10
+            ((0, 1, 0), (-1, 0, 0), (0, 0, 1)),    # 14: 4^-_001
+            ((0, -1, 0), (1, 0, 0), (0, 0, 1)),    # 15: 4^+_001
+            ((1, 0, 0), (0, 0, 1), (0, -1, 0)),    # 16: 4^-_100
+            ((-1, 0, 0), (0, 0, 1), (0, 1, 0)),    # 17: 2_011
+            ((-1, 0, 0), (0, 0, -1), (0, -1, 0)),  # 18: 2_01-1
+            ((1, 0, 0), (0, 0, -1), (0, 1, 0)),    # 19: 4^+_100
+            ((0, 0, 1), (0, 1, 0), (-1, 0, 0)),    # 20: 4^+_010
+            ((0, 0, 1), (0, -1, 0), (1, 0, 0)),    # 21: 2_101
+            ((0, 0, -1), (0, 1, 0), (1, 0, 0)),    # 22: 4^-_010
+            ((0, 0, -1), (0, -1, 0), (-1, 0, 0)),  # 23: 2_-101
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 24: -1
+            ((1, 0, 0), (0, 1, 0), (0, 0, -1)),    # 25: m_001
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),    # 26: m_010
+            ((-1, 0, 0), (0, 1, 0), (0, 0, 1)),    # 27: m_100
+            ((0, 0, -1), (-1, 0, 0), (0, -1, 0)),  # 28: -3^+_111
+            ((0, 0, -1), (1, 0, 0), (0, 1, 0)),    # 29: -3^+_-11-1
+            ((0, 0, 1), (1, 0, 0), (0, -1, 0)),    # 30: -3^+_1-1-1
+            ((0, 0, 1), (-1, 0, 0), (0, 1, 0)),    # 31: -3^+_-1-11
+            ((0, -1, 0), (0, 0, -1), (-1, 0, 0)),  # 32: -3^-_111
+            ((0, 1, 0), (0, 0, -1), (1, 0, 0)),    # 33: -3^-_1-1-1
+            ((0, -1, 0), (0, 0, 1), (1, 0, 0)),    # 34: -3^-_-1-11
+            ((0, 1, 0), (0, 0, 1), (-1, 0, 0)),    # 35: -3^-_-11-1
+            ((0, -1, 0), (-1, 0, 0), (0, 0, 1)),   # 36: m_110
+            ((0, 1, 0), (1, 0, 0), (0, 0, 1)),     # 37: m_1-10
+            ((0, -1, 0), (1, 0, 0), (0, 0, -1)),   # 38: -4^-_001
+            ((0, 1, 0), (-1, 0, 0), (0, 0, -1)),   # 39: -4^+_001
+            ((-1, 0, 0), (0, 0, -1), (0, 1, 0)),   # 40: -4^-_100
+            ((1, 0, 0), (0, 0, -1), (0, -1, 0)),   # 41: m_011
+            ((1, 0, 0), (0, 0, 1), (0, 1, 0)),     # 42: m_01-1
+            ((-1, 0, 0), (0, 0, 1), (0, -1, 0)),   # 43: -4^+_100
+            ((0, 0, -1), (0, -1, 0), (1, 0, 0)),   # 44: -4^+_010
+            ((0, 0, -1), (0, 1, 0), (-1, 0, 0)),   # 45: m_101
+            ((0, 0, 1), (0, -1, 0), (-1, 0, 0)),   # 46: -4^-_010
+            ((0, 0, 1), (0, 1, 0), (1, 0, 0)),     # 47: m_-101
+        ]
+    ],
+}
+# fmt: on
+
+# Figure 3.2.1.3 of ITA (2016)
+# tuple of (symbol of normal subgroup, coset representative)
+pg_solvable_chain = {
+    "1": [],
+    "-1": [(("1", 0), 1)],  # -x,-y,-z
+    "2": [
+        # unique axis-a
+        (("1", 0), 1),  # x,-y,-z
+        # unique axis-b
+        (("1", 0), 1),  # -x,y,-z
+        # unique axis-c
+        (("1", 0), 1),  # -x,-y,z
+    ],
+    "m": [
+        # unique axis-a
+        (("1", 0), 1),  # -x,y,z
+        # unique axis-b
+        (("1", 0), 1),  # x,-y,z
+        # unique axis-c
+        (("1", 0), 1),  # x,y,-z
+    ],
+    "2/m": [
+        # unique axis-b
+        (("2", 1), 2)  # -x,-y,-z
+    ],
+    "222": [(("2", 2), 2)],  # -x,y,-z
+    "mm2": [
+        # unique axis-a
+        (("2", 0), 2),  # x,y,-z
+        # unique axis-b
+        (("2", 1), 2),  # x,y,-z
+        # unique axis-c
+        (("2", 2), 2),  # x,y,-z
+    ],
+    "mmm": [(("222", 0), 4)],  # -x,-y,-z
+    "4": [(("1", 0), 2)],  # -y,x,z  # Never decompose "4" to "2"!
+    "-4": [(("1", 0), 2)],  # y,-x,-z  # Never decompose "-4" to "2"!
+    "4/m": [(("4", 0), 4)],  # -x,-y,-z
+    "422": [(("4", 0), 4)],  # -x,y,-z
+    "4mm": [(("4", 0), 4)],  # x,-y,z
+    "-42m": [
+        # -42m
+        (("-4", 0), 4),  # -x,y,-z
+        # -4m2
+        (("-4", 0), 4),  # x,-y,z
+    ],
+    "4/mmm": [(("422", 0), 8)],  # -x,-y,-z
+    "3": [(("1", 0), 1)],  # -y,x-y,z
+    "-3": [(("3", 0), 3)],  # -x,-y,-z
+    "32": [
+        # 312
+        (("3", 0), 3),  # -y,-x,z
+        # 321
+        (("3", 0), 3),  # y,x,-z
+    ],
+    "3m": [
+        # 3m1
+        (("3", 0), 3),  # -y,-x,z
+        # 31m
+        (("3", 0), 3),  # y,x,z
+    ],
+    "-3m": [
+        # -31m
+        (("32", 0), 6),  # -x,-y,-z
+        # -3m1
+        (("32", 1), 6),  # -x,-y,-z
+    ],
+    "6": [(("3", 0), 3)],  # -x,-y,z
+    "-6": [(("3", 0), 3)],  # x,y,-z
+    "6/m": [(("6", 0), 6)],  # -x,-y,-z
+    "622": [(("6", 0), 6)],  # y,x,-z
+    "6mm": [(("6", 0), 6)],  # -y,-x,z
+    "-6m2": [
+        # -6m2
+        (("-6", 0), 6),  # -y,-x,z
+        # -62m
+        (("-6", 0), 6),  # y,x,-z
+    ],
+    "6/mmm": [(("622", 0), 12)],  # -x,-y,-z
+    "23": [(("222", 0), 4)],  # z,x,y
+    "m-3": [(("23", 0), 12)],  # -x,-y,-z
+    "432": [(("23", 0), 12)],  # y,x,-z
+    "-43m": [(("23", 0), 12)],  # y,x,z
+    "m-3m": [(("432", 0), 24)],  # -x,-y,-z
+}
+
+
+def get_generators(pg_symbol, idx):
+    """Return indices of symmetry operations which generate given point group."""
+    _, _, gens = _get_generators(pg_symbol, idx, [])
+    return gens
+
+
+def _get_generators(pg_symbol, idx, gens):
+    if pg_symbol == "1":
+        return None, None, gens
+
+    (next_pg_symbol, next_idx), generator = pg_solvable_chain[pg_symbol][idx]
+    gens.append(generator)
+    # assert pg_dataset[pg_symbol][idx][:generator] == pg_dataset[next_pg_symbol][next_idx]
+    return _get_generators(next_pg_symbol, next_idx, gens)
+
+
+def get_pointgroup_chain_generators(prim_rotations: NDArrayInt) -> list[int]:
+    r"""Calculate generators of given crystallographic point group in primitive basis.
+
+    The returned generators give a normal series whose factor groups are all Abelian.
+
+    Parameters
+    ----------
+    prim_rotations: (order, 3, 3)
+
+    Returns
+    -------
+    generators: list[int]
+        Let :math:`G_{0} := G` and :math:`G_{i} := G_{i-1} / \langle` ``solvable_chain_generators[i]`` :math:`\rangle` (i = 0, 1, ...).
+        Then, :math:`G_{i}` is normal subgroup of :math:`G_{i-1}` and factor group :math:`G_{i-1}/G_{i}` is Abelian.
+    """
+    pg_symbol, _, P = get_pointgroup(prim_rotations)
+    Pinv = np.linalg.inv(P)
+
+    # Match given crystallographic point group with standardized ones in primitive basis.
+    order = len(prim_rotations)
+    for idx, std_rotations in enumerate(pg_dataset[pg_symbol]):
+        matched = [ndarray2d_to_integer_tuple(Pinv @ r @ P) for r in prim_rotations]
+
+        success = True
+        mapping = [-1 for _ in range(order)]  # s.t. prim_rotations[mapping[i]] == std_rotations[i]
+        for i, ri in enumerate(std_rotations):
+            try:
+                j = matched.index(ri)  # type: ignore
+            except ValueError:
+                success = False
+                break
+            mapping[i] = j
+        if success:
+            generators = get_generators(pg_symbol, idx)
+            return [mapping[g] for g in generators]
+
+    ValueError("Failed to match with tabulated point groups.")
+    return []  # This line is unreachable, just for persuading mypy!
